// src/api/models/questionSetInput.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * GradeFlow API
 * Thin FastAPI wrapper over GradeFlow Engine
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionSetInputQuestionMap } from './questionSetInputQuestionMap';

export interface QuestionSetInput {
  question_map: QuestionSetInputQuestionMap;
}