// src/api/models/choiceQuestion.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * GradeFlow API
 * Thin FastAPI wrapper over GradeFlow Engine
 * OpenAPI spec version: 0.1.0
 */
import type { ChoiceQuestionDescription } from './choiceQuestionDescription';
import type { MultiValuedParserConfig } from './multiValuedParserConfig';

export interface ChoiceQuestion {
  /** Maximum points for the question */
  max_points?: number;
  /** Optional description of the question. */
  description?: ChoiceQuestionDescription;
  type?: 'CHOICE';
  /** Parser configuration for choice questions. */
  config?: MultiValuedParserConfig;
  /** Set of valid choices for this question. */
  options?: string[];
  /** Whether to allow multiple choices to be selected. */
  allow_multiple?: boolean;
}