// src/api/models/multiValuedQuestion.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * GradeFlow API
 * Thin FastAPI wrapper over GradeFlow Engine
 * OpenAPI spec version: 0.1.0
 */
import type { MultiValuedQuestionDescription } from './multiValuedQuestionDescription';
import type { MultiValuedParserConfig } from './multiValuedParserConfig';
import type { MultiValuedQuestionValueTypesItem } from './multiValuedQuestionValueTypesItem';

export interface MultiValuedQuestion {
  /** Maximum points for the question */
  max_points?: number;
  /** Optional description of the question. */
  description?: MultiValuedQuestionDescription;
  type?: 'MULTI_VALUED';
  /** Parser configuration for multi-valued questions. */
  config?: MultiValuedParserConfig;
  /** Expected type for each value in the answer. */
  value_types: MultiValuedQuestionValueTypesItem[];
}