// src/api/models/textQuestion.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * GradeFlow API
 * Thin FastAPI wrapper over GradeFlow Engine
 * OpenAPI spec version: 0.1.0
 */
import type { TextQuestionDescription } from './textQuestionDescription';
import type { TextParserConfig } from './textParserConfig';

export interface TextQuestion {
  /** Maximum points for the question */
  max_points?: number;
  /** Optional description of the question. */
  description?: TextQuestionDescription;
  type?: 'TEXT';
  /** Parser configuration for text questions. */
  config?: TextParserConfig;
}