// src/api/models/numericQuestion.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * GradeFlow API
 * Thin FastAPI wrapper over GradeFlow Engine
 * OpenAPI spec version: 0.1.0
 */
import type { NumericQuestionDescription } from './numericQuestionDescription';
import type { BaseParserConfig } from './baseParserConfig';

export interface NumericQuestion {
  /** Maximum points for the question */
  max_points?: number;
  /** Optional description of the question. */
  description?: NumericQuestionDescription;
  type?: 'NUMERIC';
  /** Parser configuration for numeric questions. */
  config?: BaseParserConfig;
}